requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
requests==2.25.1
python-dotenv==0.17.1
flask-swagger-ui==3.36.0